Use all references, not just those where the condition is true
This is useful when you have a IsRunningOnBuildServer condition or many different build configurations but want to just convert everything.
This could cause problems if people wanted to use subtly different packages depending on condition tags, but NuGet tends to nuke those conditions every update anyway so it'd be pretty tricky to get such a system working.