Convert project references too
This could be problematic for those previously using a repo containing the output of a multi-project build but I've found it's always best to create a local folder containing the desired packages anyway, and that is indeed what the readme suggests.
If this becomes an issue for people, explicitly mentioning this in the readme might be enough to help people avoid the issue.